Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq

faiss-cpu
fastembed
pypdf

firebase-admin
redis

python-dotenv
httpx
tqdm
numpy
